Add initial implementations and test/benchmark runners
This commit adds a basic test runner for the various implementations.
This doesn't use JUnit, because I didn't want to complicate things, and
we know that the original implementation is not thread-safe, so would
break the build.

There's also a JMH-based benchmark runner, and a script to run it with
various number of threads.

The implementations are:

* OriginalCounter: The code as presented in the original article
* MapComputeCounter: This uses Map.compute() with MutableInteger
* AtomicIntegerCounter: This uses AtomicInteger in a ConcurrentHashMap
* LongAdderCounter: This uses LongAdder inside a ConcurrentHashMap
* InteritanceAtomicIntegerCounter: Modified AtomicIntegerCounter to use
        inheritance instead of composition to see if there's any perform
        differences.
* SynchronizedHashMapCounter: Modified AtomicIntegerCounter to use
        Collections.synchronizedMap() with a HashMap to see the differences
        in performance with a ConcurrentHashMap.