Add a non-Concurrent Map version of AtomicInteger
I was curious to see what effect we get from using a normal HashMap
instead of a ConcurrentHashMap.

This version is not thread-safe. At best, we lose several counts at the
start, but no more than the number of threads running when a new key is
added. At worse, it throws a ConcurrentModificationException.

Spoiler: This version is slower than the ConcurrentHashMap version.